+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+###20170915
+
+*   配置 xpath 提取目标数据
+```aidl
+    目标 url
+    https://kuai.baidu.com/webapp/bus/list.html?startcityid=%257B%2522cityid%2522%253A131%252C%2522cityname%2522%253A%2522%25E5%258C%2597%25E4%25BA%25AC%2522%252C%2522type%2522%253A2%257D&arrivalcityid=%257B%2522cityid%2522%253A307%252C%2522cityname%2522%253A%2522%25E4%25BF%259D%25E5%25AE%259A%2522%252C%2522type%2522%253A2%257D&startname=北京&arrivename=保定&startdatetime=1505471400000
+
+    <property>
+        <name>xpath-kuai.baidu.com-/webapp/bus/list.html</name>
+        <value>//div[@class='js_list_wrapper']/ul</value>
+        <description>
+            System.out.println(url.getHost()); kuai.baidu.com/
+            System.out.println(url.getPath()); webapp/bus/list.html
+
+        </description>
+    </property>
+    
+
+```
+
 ###20170913
 
 *   json 数据的 ETL 处理： json to csv